--- backend/internal/service/handler/reviews/vote.go
@@ -1,59 +1,0 @@
-package vote
-
-import (
-	"github.com/gofiber/fiber/v2"
-	"platnm/internal/models"
-	"platnm/internal/errs"
-)
-
-type createVoteRequest struct {
-	models.UserReviewVote
-}
-
-func (h *Handler) VoteReview(c *fiber.Ctx) error {
-	var req createVoteRequest
-	if err := c.BodyParser(&req); err != nil {
-		return errs.BadRequest("failed to parse request body")
-	}
-	exists, err := h.userRepository.UserExists(c.Context(), req.UserReviewVote.UserID)
-	if err != nil {
-		return err
-	}
-
-	if !exists {
-		return errs.NotFound("User", "id", &req.UserReviewVote.UserID)
-	}
-
-	reviewExists, reviewExistsErr := h.reviewRepository.ReviewExists(c.Context(),req.UserReviewVote.ReviewID)
-	if reviewExistsErr != nil {
-		return reviewExistsErr
-	}
-
-	if reviewExists == false {
-		return errs.NotFound("Review", "id", &req.UserReviewVote.ReviewID)
-	}
-
-	voteValue, voteExistErr := h.voteRepository.GetVoteIfExists(c.Context(), req.UserReviewVote.UserID, req.UserReviewVote.ReviewID)
-	if voteExistErr != nil {
-		return voteExistErr
-	}
-
-	if voteValue != nil {
-		voteErr := h.voteRepository.AddVote(c.Context(), &req.UserReviewVote)
-		if voteErr != nil {
-			return voteErr
-		}
-	} else if req.UserReviewVote.Upvote == voteValue.Upvote {
-		delVoteErr := h.voteRepository.DeleteVote(c.Context(), req.UserReviewVote.UserID, req.UserReviewVote.ReviewID)
-		if delVoteErr != nil {
-			return delVoteErr
-		}
-	} else { 
-		updateVoteErr := h.voteRepository.UpdateVote(c.Context(), req.UserReviewVote.UserID, req.UserReviewVote.ReviewID, req.UserReviewVote.Upvote)
-		if updateVoteErr != nil {
-			return updateVoteErr
-		}
-	}
-
-	return c.Status(fiber.StatusOK).JSON(voteValue)
-}
